refactor: rename app to storage
