+++ nav-bar-sticky/src/App.tsx
@@ -1,4 +1,4 @@
-import { useAppSelector } from './app/hooks';
+import { useAppSelector } from './storage/hooks';
 import { displayPage } from './components/nav-bar/pageSlice';
 import styled from 'styled-components';
 import {
