+++ nav-bar-sticky/src/components/nav-bar/pageSlice.tsx
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
-import type { RootState } from '../../app/store';
+import type { RootState } from '../../storage/store';
 
 interface holderState {
   value: string;
